Exercises/Exercise11/setup/requirements.txt
pymongo==4.15.4
python_dateutil==2.9.0.post0
